[#35][refactor] 컨트롤러, 서비스 코드 분리
기존에 상품과 관심상품, 상태에 관련된 productController와 ProductService 코드를 Product와 Favorite, AuctionStatus로  개념을 나누어 각각의 컨트롤러, 서비스를 만들었음.
