+++ src/main/java/f4/product/domain/product/service/ProductImageService.java
@@ -1,0 +1,5 @@
+package f4.product.domain.product.service;
+
+public interface ProductImageService {
+
+}
